+++ tests/BillingReferenceCreditNoteTest.php
@@ -5,11 +5,11 @@ namespace NumNum\UBL\Tests;
 use PHPUnit\Framework\TestCase;
 
 /**
- * Test an UBL2.1 invoice document
+ * Test an UBL2.1 credit note document
  */
-class SimpleInvoiceTest extends TestCase
+class BillingReferenceCreditNoteTest extends TestCase
 {
-    private $schema = 'http://docs.oasis-open.org/ubl/os-UBL-2.1/xsd/maindoc/UBL-Invoice-2.1.xsd';
+    private $schema = 'http://docs.oasis-open.org/ubl/os-UBL-2.1/xsd/maindoc/UBL-CreditNote-2.1.xsd';
 
     /** @test */
     public function testIfXMLIsValid()
@@ -32,18 +32,10 @@ class SimpleInvoiceTest extends TestCase
             ->setPhysicalLocation($address)
             ->setPostalAddress($address);
 
-        // Client contact node
-        $clientContact = (new \NumNum\UBL\Contact())
-            ->setName('Client name')
-            ->setElectronicMail('email@client.com')
-            ->setTelephone('0032 472 123 456')
-            ->setTelefax('0032 9 1234 567');
-
         // Client company node
         $clientCompany = (new \NumNum\UBL\Party())
             ->setName('My client')
-            ->setPostalAddress($address)
-            ->setContact($clientContact);
+            ->setPostalAddress($address);
 
         $legalMonetaryTotal = (new \NumNum\UBL\LegalMonetaryTotal())
             ->setPayableAmount(10 + 2)
@@ -57,7 +49,8 @@ class SimpleInvoiceTest extends TestCase
         $productItem = (new \NumNum\UBL\Item())
             ->setName('Product Name')
             ->setDescription('Product Description')
-            ->setSellersItemIdentification('SELLERID');
+            ->setSellersItemIdentification('SELLERID')
+            ->setBuyersItemIdentification('BUYERID');
 
         // Price
         $price = (new \NumNum\UBL\Price())
@@ -69,45 +62,15 @@ class SimpleInvoiceTest extends TestCase
         $lineTaxTotal = (new \NumNum\UBL\TaxTotal())
             ->setTaxAmount(2.1);
 
-        // InvoicePeriod
-        $invoicePeriod = (new \NumNum\UBL\InvoicePeriod())
-            ->setStartDate(new \DateTime());
-
         // Invoice Line(s)
-        $invoiceLines = [];
-
-        $orderLineReference = (new \NumNum\UBL\OrderLineReference)
-            ->setLineId('#ABC123');
-
-        $invoiceLines[] = (new \NumNum\UBL\InvoiceLine())
+        $creditNoteLine = (new \NumNum\UBL\CreditNoteLine())
             ->setId(0)
             ->setItem($productItem)
-            ->setInvoicePeriod($invoicePeriod)
             ->setPrice($price)
             ->setTaxTotal($lineTaxTotal)
-            ->setInvoicedQuantity(1)
-            ->setOrderLineReference($orderLineReference);
+            ->setInvoicedQuantity(1);
 
-        $invoiceLines[] = (new \NumNum\UBL\InvoiceLine())
-            ->setId(0)
-            ->setItem($productItem)
-            ->setInvoicePeriod($invoicePeriod)
-            ->setPrice($price)
-            ->setAccountingCost('Product 123')
-            ->setTaxTotal($lineTaxTotal)
-            ->setInvoicedQuantity(1)
-            ->setOrderLineReference($orderLineReference);
-
-        $invoiceLines[] = (new \NumNum\UBL\InvoiceLine())
-            ->setId(0)
-            ->setItem($productItem)
-            ->setInvoicePeriod($invoicePeriod)
-            ->setPrice($price)
-            ->setAccountingCostCode('Product 123')
-            ->setTaxTotal($lineTaxTotal)
-            ->setInvoicedQuantity(1)
-            ->setOrderLineReference($orderLineReference);
-
+        $creditNoteLines = [$creditNoteLine];
 
         // Total Taxes
         $taxCategory = (new \NumNum\UBL\TaxCategory())
@@ -121,34 +84,40 @@ class SimpleInvoiceTest extends TestCase
             ->setTaxAmount(2.1)
             ->setTaxCategory($taxCategory);
 
-
         $taxTotal = (new \NumNum\UBL\TaxTotal())
             ->addTaxSubTotal($taxSubTotal)
             ->setTaxAmount(2.1);
 
+        $billingReference = (new \NumNum\UBL\BillingReference())
+            ->setInvoiceDocumentReference((new \NumNum\UBL\InvoiceDocumentReference())
+                    ->setOriginalInvoiceId(1234)
+                    ->setIssueDate(new \DateTime()));
+
+
         // Invoice object
-        $invoice = (new \NumNum\UBL\Invoice())
+        $creditNote = (new \NumNum\UBL\CreditNote())
             ->setId(1234)
             ->setCopyIndicator(false)
             ->setIssueDate(new \DateTime())
             ->setAccountingSupplierParty($supplierCompany)
             ->setAccountingCustomerParty($clientCompany)
-            ->setSupplierAssignedAccountID('10001')
-            ->setInvoiceLines($invoiceLines)
+            ->setBillingReference($billingReference)
+            ->setCreditNoteLines($creditNoteLines)
             ->setLegalMonetaryTotal($legalMonetaryTotal)
-            ->setTaxTotal($taxTotal);
+            ->setTaxTotal($taxTotal)
+            ->setInvoiceTypeCode(\NumNum\UBL\InvoiceTypeCode::CREDIT_NOTE);
 
         // Test created object
         // Use \NumNum\UBL\Generator to generate an XML string
         $generator = new \NumNum\UBL\Generator();
-        $outputXMLString = $generator->invoice($invoice);
+        $outputXMLString = $generator->creditNote($creditNote);
 
         // Create PHP Native DomDocument object, that can be
         // used to validate the generate XML
         $dom = new \DOMDocument;
         $dom->loadXML($outputXMLString);
 
-        $dom->save('./tests/SimpleInvoiceTest.xml');
+        $dom->save('./tests/SimpleCreditNoteTest.xml');
 
         $this->assertEquals(true, $dom->schemaValidate($this->schema));
     }