+++ tests/SimpleInvoiceWithInlinePdfTest.php
@@ -7,7 +7,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * Test an UBL2.1 invoice document
  */
-class SimpleInvoiceTest extends TestCase
+class SimpleInvoiceWithInlinePdfTest extends TestCase
 {
     private $schema = 'http://docs.oasis-open.org/ubl/os-UBL-2.1/xsd/maindoc/UBL-Invoice-2.1.xsd';
 
@@ -126,9 +126,21 @@ class SimpleInvoiceTest extends TestCase
             ->addTaxSubTotal($taxSubTotal)
             ->setTaxAmount(2.1);
 
+        // Example if you have some inline or in-memory filestream/file contents
+        $fileStream = file_get_contents(__DIR__.DIRECTORY_SEPARATOR.'SampleInvoice.pdf'); // this would be your file contents
+
+        $base64EncodedFileStream = base64_encode($fileStream);
+
+        $attachment = (new \NumNum\UBL\Attachment())
+            ->setFileStream($base64EncodedFileStream, 'Invoice.pdf', 'application/pdf');
+
+        $additionalDocumentReference = (new \NumNum\UBL\AdditionalDocumentReference())
+            ->setAttachment($attachment);
+
         // Invoice object
         $invoice = (new \NumNum\UBL\Invoice())
             ->setId(1234)
+            ->setAdditionalDocumentReference($additionalDocumentReference)
             ->setCopyIndicator(false)
             ->setIssueDate(new \DateTime())
             ->setAccountingSupplierParty($supplierCompany)
@@ -148,7 +160,7 @@ class SimpleInvoiceTest extends TestCase
         $dom = new \DOMDocument;
         $dom->loadXML($outputXMLString);
 
-        $dom->save('./tests/SimpleInvoiceTest.xml');
+        $dom->save('./tests/SimpleInvoiceWithInlinePdfTest.xml');
 
         $this->assertEquals(true, $dom->schemaValidate($this->schema));
     }